+++ modules/AWSPowerShell/Cmdlets/Backup/Basic/Get-BAKRestoreJobList-Cmdlet.cs
@@ -52,6 +52,28 @@ namespace Amazon.PowerShell.Cmdlets.BAK
         public System.String ByAccountId { get; set; }
         #endregion
         
+        #region Parameter ByCompleteAfter
+        /// <summary>
+        /// <para>
+        /// <para>Returns only copy jobs completed after a date expressed in Unix format and Coordinated
+        /// Universal Time (UTC).</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public System.DateTime? ByCompleteAfter { get; set; }
+        #endregion
+        
+        #region Parameter ByCompleteBefore
+        /// <summary>
+        /// <para>
+        /// <para>Returns only copy jobs completed before a date expressed in Unix format and Coordinated
+        /// Universal Time (UTC).</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public System.DateTime? ByCompleteBefore { get; set; }
+        #endregion
+        
         #region Parameter ByCreatedAfter
         /// <summary>
         /// <para>
@@ -152,6 +174,8 @@ namespace Amazon.PowerShell.Cmdlets.BAK
                     throw new System.ArgumentException("Invalid value for -Select parameter.", nameof(this.Select));
             }
             context.ByAccountId = this.ByAccountId;
+            context.ByCompleteAfter = this.ByCompleteAfter;
+            context.ByCompleteBefore = this.ByCompleteBefore;
             context.ByCreatedAfter = this.ByCreatedAfter;
             context.ByCreatedBefore = this.ByCreatedBefore;
             context.ByStatus = this.ByStatus;
@@ -188,6 +212,14 @@ namespace Amazon.PowerShell.Cmdlets.BAK
             if (cmdletContext.ByAccountId != null)
             {
                 request.ByAccountId = cmdletContext.ByAccountId;
+            }
+            if (cmdletContext.ByCompleteAfter != null)
+            {
+                request.ByCompleteAfter = cmdletContext.ByCompleteAfter.Value;
+            }
+            if (cmdletContext.ByCompleteBefore != null)
+            {
+                request.ByCompleteBefore = cmdletContext.ByCompleteBefore.Value;
             }
             if (cmdletContext.ByCreatedAfter != null)
             {
@@ -263,6 +295,14 @@ namespace Amazon.PowerShell.Cmdlets.BAK
             if (cmdletContext.ByAccountId != null)
             {
                 request.ByAccountId = cmdletContext.ByAccountId;
+            }
+            if (cmdletContext.ByCompleteAfter != null)
+            {
+                request.ByCompleteAfter = cmdletContext.ByCompleteAfter.Value;
+            }
+            if (cmdletContext.ByCompleteBefore != null)
+            {
+                request.ByCompleteBefore = cmdletContext.ByCompleteBefore.Value;
             }
             if (cmdletContext.ByCreatedAfter != null)
             {
@@ -396,6 +436,8 @@ namespace Amazon.PowerShell.Cmdlets.BAK
         internal partial class CmdletContext : ExecutorContext
         {
             public System.String ByAccountId { get; set; }
+            public System.DateTime? ByCompleteAfter { get; set; }
+            public System.DateTime? ByCompleteBefore { get; set; }
             public System.DateTime? ByCreatedAfter { get; set; }
             public System.DateTime? ByCreatedBefore { get; set; }
             public Amazon.Backup.RestoreJobStatus ByStatus { get; set; }